% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleUnitConverters.R
\name{cm2in}
\alias{cm2in}
\title{Convert centimeter to inch}
\usage{
cm2in(cm)
}
\arguments{
\item{cm}{numeric. Centimeter to be converted to inch}
}
\value{
numeric. Given centimeter in inch.
}
\description{
Simple converter from centimeter to inch.
}
\examples{
  # One inch is defined as 25.4 millimeter
  cm2in(2.54)

  # Works with a vector too
  cm2in(1:10)

}
\seealso{
Inverse conversion \code{\link[=in2cm]{in2cm()}}.
}
