% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countElementsInFormular.R
\name{countElementsInFormula}
\alias{countElementsInFormula}
\title{Counts the elements a chemical formula}
\usage{
countElementsInFormula(fmla)
}
\arguments{
\item{fmla}{Formula to count the elements therein as a character vector -- as
string.}
}
\value{
A numeric vector with the found elements as names and their number of
occurrence. This vector is ordered by the names of the elements
}
\description{
Counts the elements occurring in a given chemical formula. The formula is
split into their elements using a regular expression.
}
\examples{
  countElementsInFormula('Mg2(SiO4)')
  countElementsInFormula('CaMgSi2O6')
  countElementsInFormula('Mg3Al2(SiO4)3')
  countElementsInFormula('NaMg3Al6(Si6O18)(BO3)3(OH)3(OH)')

}
\references{
The regular expression used herein was found on
\url{https://regex101.com/r/vE6nC7/1}
}
