% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMetrix.R
\name{dist.chebyshev}
\alias{dist.chebyshev}
\title{Chebyshev distance between two vectors}
\usage{
dist.chebyshev(X, Y)
}
\arguments{
\item{X}{numeric vector. Coordinates of one point.}

\item{Y}{numeric vector. Coordinates of the other point.}
}
\value{
numeric scalar. The Chebyshev distance between the two vectors.
}
\description{
The Chebyshev distance is the maximum of the absolute differences of the
coordinates of two vectors.
}
\examples{
# Classic example: How many moves must a king make on a chessboard to get
# from e.g. `f3` (6, 3) to `b8` (2, 8)?
dist.chebyshev(c(6, 3), c(2, 8))

}
