% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMetrix.R
\name{dist.euclidean}
\alias{dist.euclidean}
\title{Euclidean distance between two points}
\usage{
dist.euclidean(X, Y)
}
\arguments{
\item{X}{numeric vector. Coordinates of one point.}

\item{Y}{numeric vector. Coordinates of the other point.}
}
\value{
numeric scalar. The euclidean distance between the two given points.
}
\description{
Calculate the Euclidean distance between two points. The Euclidean distance
is the distance between two points if they are connected by a straight line.
It is the square root of the sum of differences between two points.
}
\examples{
# Length of straight line in two dimensions -- equal `sqrt(2)`
dist.euclidean(c(0, 0), c(1, 1))

# Length of straight line in three dimensions -- equal `sqrt(3)`
dist.euclidean(c(0, 0, 0), c(1, 1, 1))

}
