% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMetrix.R
\name{dist.manhattan}
\alias{dist.manhattan}
\title{Manhattan distance between two points}
\usage{
dist.manhattan(X, Y)
}
\arguments{
\item{X}{numeric vector. Coordinates of one point.}

\item{Y}{numeric vector. Coordinates of the other point.}
}
\value{
numeric scalar. The Manhattan distance between the two given points.
}
\description{
Calculate the Manhattan distance also known as the block distance or taxicab
distance between two points. It is the sum of absolute differences of the
(Cartesian) coordinates of two different points.
}
\examples{
# Distance to walk on the diagonal corner of the unit block
dist.manhattan(c(0, 0), c(1, 1))

# Distance to walk on the diagonal corner of the unit block and one floor up
dist.manhattan(c(0, 0, 0), c(1, 1, 1))

}
