% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_oxide.R
\name{eloxFactor}
\alias{eloxFactor}
\title{Element to stoichiometric oxide conversion factor}
\usage{
eloxFactor(oxide)
}
\arguments{
\item{oxide}{character. The chemical formula of an oxide the conversion factor shall be calculated for}
}
\value{
numeric. The conversion factor.
}
\description{
A general function to calculate the factor needed to convert an element
concentration to it's stoichiometric oxide or vice versa.\cr
To convert an oxide concentration to it's element concentration multiply by this factor. To convert an element concentration to the equivalent expressed as oxide divide by this factor.
}
\examples{
# convert 0.25 lithium (\%wt) to it's equivalent expressed as Li2O
0.25 / eloxFactor("Li2O")

}
\seealso{
\itemize{
\item \code{\link[=el2ox]{el2ox()}}, \code{\link[=ox2el]{ox2el()}} for conversion functions
\item \link{elOxTable}, \link{elOxTable_121} lookup tables for the conversation
\item \code{\link[=mass]{mass()}} to calculate the molecular mass
\item \link[georefdatar:IUPAC_StdAW]{georefdatar::IUPAC_StdAW} and \code{\link[georefdatar:aw]{georefdatar::aw()}} for the standard
atomic weights.
\item \link[georefdatar:pte]{georefdatar::pte} for a list of oxidation states of an element
}
}
