% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_oxide.R
\docType{data}
\name{elOxTable}
\alias{elOxTable}
\alias{elOxTable_121}
\title{Table of Elements, oxides and their conversion factors}
\format{
A data frame with the following columns:\cr
\describe{
\item{Element}{character, Element's symbol}
\item{Oxide}{character, Element oxide. Depending on the possible oxidation states of an element, there can be a one-to-many relationship between the element and its oxides (\code{elOxTable}vs \code{elOxTable_121}). These are the polyvalent oxides e.g. Mn, V, U, ...}
\item{Factor}{numeric, Conversion factor to convert a concentration of an element to its oxide or vice versa.}
}

An object of class \code{data.frame} with 77 rows and 3 columns.
}
\usage{
elOxTable

elOxTable_121
}
\description{
A simple lookup table (LUT) of elements and their corresponding oxides, and
the conversion factor to convert a concentration given in oxide to an element
concentration or vice versa.\cr The oxides where built using the positive
oxidation states of an element as given in the periodic table of elements in
the georefdatar package (\link[georefdatar:pte]{georefdatar::pte}). The conversion factors were
then calculated using the standard atomic weights of the elements recommended
by the IUPAC and are quoted to five significant figures. So here are those
elements listed for those a standard atomic weight is given by the IUPAC.\cr
These factors are calculated and stored here for convenience and speed. They
could actually be calculated on the fly during a conversion.
}
\details{
\itemize{
\item \code{elOxTable} holds the full of elements and oxides. There will be a
one-to-many relation between the elements, having more than one oxidation state, and their oxides -- e.g. Fe2+, Fe3+, Mn, Ti, U, ...
\item \code{elOxTable_121} is a subset of \code{elOxTable} which holds only one (common) oxide for an element -- a one-to-one relationship between element and oxide. This is convenience and makes the use of \code{\link[=el2ox]{el2ox()}} easier, since there is no need to pass an oxide to this function if one wants to limit the result to one.
}
}
\seealso{
\itemize{
\item \code{\link[=el2ox]{el2ox()}}, \code{\link[=ox2el]{ox2el()}} for conversion functions
\item \code{\link[=eloxFactor]{eloxFactor()}} for the conversion factor
\item \link[georefdatar:pte]{georefdatar::pte} for a list of oxidation states of an element
\item \link[georefdatar:IUPAC_StdAW]{georefdatar::IUPAC_StdAW} and \code{\link[georefdatar:aw]{georefdatar::aw()}} for the standard
atomic weights.
}
}
\keyword{datasets}
