% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hularr-package.R
\docType{package}
\name{hularr-package}
\alias{hularr}
\alias{hularr-package}
\title{hularr: Happily Used Little R Routines}
\description{
This is a hodgepodge of regularly and happily used little R routines. They are small tools for recurring common tasks that are collected here.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abuseki/hularr}
  \item Report bugs at \url{https://github.com/abuseki/hularr/issues}
}

}
\author{
\strong{Maintainer}: Gerald Schuberth-HlavaÄ \email{abuseki@synapticgap.com}

}
\keyword{internal}
