% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleUnitConverters.R
\name{in2cm}
\alias{in2cm}
\title{Convert inch to centimeter}
\usage{
in2cm(inch)
}
\arguments{
\item{inch}{numeric. Inch to be converted to centimeter.}
}
\value{
numeric. Given inch in centimeter.
}
\description{
Simple converter from inch to centimeter.
}
\examples{
  # One inch is defined as 25.4 millimeter
  in2cm(1)

  # Works with a vector too
  in2cm(1:10)

}
\seealso{
Inverse conversion \code{\link[=cm2in]{cm2in()}}
}
