% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleUnitConverters.R
\name{ppm2wt}
\alias{ppm2wt}
\title{Convert ppm to mass percentage}
\usage{
ppm2wt(ppm)
}
\arguments{
\item{ppm}{numeric. concentration in ppm to be converted to wt\%.}
}
\value{
numeric. wt\% of given ppm
}
\description{
Simple converter from ppm to wt\%
}
\examples{
  ppm2wt(10001)
  ppm2wt(10000.123:1000)

}
\seealso{
Inverse conversion \code{\link[=wt2ppm]{wt2ppm()}}
}
