% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleUnitConverters.R
\name{wt2ppm}
\alias{wt2ppm}
\title{Convert mass percentage to ppm}
\usage{
wt2ppm(wt)
}
\arguments{
\item{wt}{numeric. mass percentage to be converted to ppm.}
}
\value{
numeric. ppm of given wt\%
}
\description{
Simple converter from wt\% to ppm.
}
\examples{
  wt2ppm(56.6)
  wt2ppm(3.5:6.5)

}
\seealso{
Inverse conversion \code{\link[=ppm2wt]{ppm2wt()}}
}
