% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConverters.R
\name{ppm2wt}
\alias{ppm2wt}
\title{Convert ppm to mass percentage}
\usage{
ppm2wt(ppm)
}
\arguments{
\item{ppm}{concentration in ppm to be converted to mass percentage (wt\%) as a numeric vector}
}
\value{
numeric vector of converted ppm
}
\description{
Simple converter from ppm to wt\%
}
\examples{
  ppm2wt(10001)
  ppm2wt(10000.123:1000)

}
