% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConverters.R
\name{wt2ppm}
\alias{wt2ppm}
\title{Convert mass percentage to ppm}
\usage{
wt2ppm(wt)
}
\arguments{
\item{wt}{mass percentage (wt\%) to be converted to ppm as a numeric vector.}
}
\value{
numeric vector of converted wt\%
}
\description{
Simple converter from wt\% to ppm.
}
\examples{
  wt2ppm(56.6)
  wt2ppm(3.5:6.5)

}
